OAK-4868: Update SegmentS3DataStoreBlobGCTest in oak-it once oak-segment-tar 0.0.18 released
* Removed calls to super clean up methods as already tagged with @After
* Checked for bucket null or empty before cleaning up

git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1766495 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-it/src/test/java/org/apache/jackrabbit/oak/segment/SegmentS3DataStoreBlobGCIT.java
@@ -28,6 +28,7 @@ import java.util.Date;
 import java.util.List;
 import java.util.Properties;
 
+import com.google.common.base.Strings;
 import org.apache.jackrabbit.oak.blob.cloud.s3.S3Constants;
 import org.apache.jackrabbit.oak.blob.cloud.s3.S3DataStoreUtils;
 import org.apache.jackrabbit.oak.plugins.blob.datastore.DataStoreBlobStore;
@@ -75,9 +76,9 @@ public class SegmentS3DataStoreBlobGCIT extends SegmentDataStoreBlobGCIT {
 
     @After
     public void close() throws Exception {
-        super.closeBlobStore();
-        super.closeBlobStore();
-        S3DataStoreUtils.deleteBucket(bucket, startDate);
+        if (!Strings.isNullOrEmpty(bucket)) {
+            S3DataStoreUtils.deleteBucket(bucket, startDate);
+        }
     }
 }
